+++ autoload.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 /**
  *   __ _  ___  ___ ___   ___   ___     ____ _ __ ___   ___
  *  / _` |/  / / __/ _ \ / _ \ /  /    / __/| '_ ` _ \ /  /
@@ -16,57 +16,18 @@
  * @ASCOOS-COPYRIGHT   : Copyright (c) 2007 - 2023, AlexSoft Software.               *
  *************************************************************************************
  *
- * @package            : ASCOOS CMS - phpBCL
- * @subpackage         : Example mb_str_split Function
- * @source             : /phpBCL/test/mb_str_split.php
+ * @package            : phpBCL
+ * @subpackage         : Compatibilities Autoload file
+ * @source             : /phpBCL/autoload.php
  * @version            : 1.1.3
- * @created            : 2023-07-07 07:00:00 UTC+3
- * @updated            : 2024-10-22 07:00:00 UTC+3
+ * @created            : 2024-10-22 07:00:00 UTC+3
+ * @updated            : 
  * @author             : Drogidis Christos
  * @authorSite         : www.alexsoft.gr
  */
 
-/**
- * ++ 7.4.0  ---- https://www.php.net/manual/en/function.mb-str-split.php
- */
-
+ $phpBCL_path = str_replace('\\', '/', __DIR__);
  
- 
-require_once("../autoload.php");
+ include $phpBCL_path . "/src/coreCompatibilities.php";
 
-
-$str = "Hello Friend";
-
-$arr1 = mb_str_split($str);
-$arr2 = mb_str_split($str, 3);
-
-print_r($arr1);
-print_r($arr2);
-
-/*
-Array
-(
-    [0] => H
-    [1] => e
-    [2] => l
-    [3] => l
-    [4] => o
-    [5] =>
-    [6] => F
-    [7] => r
-    [8] => i
-    [9] => e
-    [10] => n
-    [11] => d
-)
-
-Array
-(
-    [0] => Hel
-    [1] => lo
-    [2] => Fri
-    [3] => end
-)
-*/
-
-?>
+ ?>