+++ src/compat/compat_php70x.php
@@ -19,16 +19,12 @@
  * @package            : ASCOOS CMS - phpBCL
  * @subpackage         : Core Compatibilities Manager for PHP < 7.0.0
  * @source             : /phpBCL/src/compat/compat_php70x.php
- * @version            : 1.0.9
+ * @version            : 1.1.3
  * @created            : 2024-02-28 07:00:00 UTC+3
- * @updated            : 
+ * @updated            : 2024-10-22 07:00:00 UTC+3
  * @author             : Drogidis Christos
  * @authorSite         : www.alexsoft.gr
  */
-
-
-// Run on ASCOOS CMS only. Marked as comment if you want run this script with other cms.
-defined ("ALEXSOFT_RUN_CMS") or die("Prohibition of Access.");
 
 
 /**