+++ src/phpBCL.php
@@ -19,16 +19,14 @@
  * @package            : ASCOOS CMS - phpBCL
  * @subpackage         : phpBCL Native Functions
  * @source             : /phpBCL/src/phpBCL.php
- * @version            : 1.0.8
+ * @version            : 1.1.3
  * @created            : 2024-02-20 05:40:00 UTC+3
- * @updated            : 2024-02-27 09:00:00 UTC+3
+ * @updated            : 2024-10-22 07:00:00 UTC+3
  * @author             : Drogidis Christos
  * @authorSite         : www.alexsoft.gr
  */
 
- 
-// Run on ASCOOS CMS only. Marked as comment if you want run this script with other cms.
-defined ("ALEXSOFT_RUN_CMS") or die("Prohibition of Access.");
+
 
 
 /**