+++ test/80__class_Stringable.php
@@ -19,9 +19,9 @@
  * @package            : ASCOOS CMS - phpBCL
  * @subpackage         : Example mb_str_split Function
  * @source             : /phpBCL/test/mb_str_split.php
- * @version            : **** - $release: 1.0 - $revision: 1 - $build: ****
+ * @version            : 1.1.3
  * @created            : 2023-07-07 07:00:00 UTC+3
- * @updated            : 
+ * @updated            : 2024-10-22 07:00:00 UTC+3
  * @author             : Drogidis Christos
  * @authorSite         : www.alexsoft.gr
  */
@@ -30,11 +30,8 @@
  * ++ 8.0.0  ---- https://www.php.net/manual/en/class.stringable.php
  */
 
- define('ALEXSOFT_RUN_CMS', true);
-
- $cms_path = str_replace('/phpBCL/test', '',str_replace('\\', '/', __DIR__));
-
- require_once($cms_path."/phpBCL/src/coreCompatibilities.php");
+ 
+ require_once("../autoload.php");
 
 
 class IPv4Address implements Stringable {
@@ -52,7 +49,7 @@ class IPv4Address implements Stringable {
 
     public function __toString()
     {
-        return "$this->oct1.$this->oct2.$this->oct3.$this->oct4";
+        return $this->oct1.'.'.$this->oct2.'.'.$this->oct3.'.'.$this->oct4;
     }
 }