+++ test/81__array_is_list.php
@@ -19,20 +19,17 @@
  * @package            : ASCOOS CMS - phpBCL
  * @subpackage         : Example mb_str_pad Function
  * @source             : /phpBCL/test/81__array_is_list.php
- * @version            : **** - $release: 1.0 - $revision: 6 - $build: ****
+ * @version            : 1.1.3
  * @created            : 2023-07-07 07:00:00 UTC+3
- * @updated            : 2023-02-21 07:00:00 UTC+3
+ * @updated            : 2024-10-22 07:00:00 UTC+3
  * @author             : Drogidis Christos
  * @authorSite         : www.alexsoft.gr
  * 
  * @since 5.6.40
  */
 
-define('ALEXSOFT_RUN_CMS', true);
-
-$cms_path = str_replace('/phpBCL/test', '', str_replace('\\', '/', __DIR__));
- 
-require_once($cms_path."/phpBCL/src/coreCompatibilities.php");
+  
+ require_once("../autoload.php");
 ?>
 <html>
 <head>