+++ test/84_array_find .php
@@ -19,20 +19,17 @@
  * @package            : ASCOOS CMS - phpBCL
  * @subpackage         : Example array_find Function
  * @source             : /phpBCL/test/84_array_find.php
- * @version            : 1.1.2
+ * @version            : 1.1.3
  * @created            : 2024-10-04 07:00:00 UTC+3
- * @updated            : 
+ * @updated            : 2024-10-22 07:00:00 UTC+3
  * @author             : Drogidis Christos
  * @authorSite         : www.alexsoft.gr
  * 
  * @since 5.6.40
  */
 
- define('ALEXSOFT_RUN_CMS', true);
 
- $cms_path = str_replace('/phpBCL/test', '',str_replace('\\', '/', __DIR__));
-  
- require_once($cms_path."/phpBCL/src/coreCompatibilities.php");
+require_once("../autoload.php");
 
 
 function is_even($value)
@@ -48,9 +45,10 @@ function is_key_numeric($value, $key)
 // Since PHP >= 5.6
 echo array_find([1, 2, 3, 4, 5], 'is_even').'<br>'; // 2  
 echo array_find(['a' => 'foo', 2 => 'bar'], 'is_key_numeric').'<br>'; // "bar"   
-
+/*
 // Since PHP >= 7.4
 if (version_compare(PHP_VERSION, '7.4.0', '>=')) { 
     echo array_find([1, 2, 3, 4, 5], fn($value) => $value % 2 === 0); // 2
 }
+*/
 ?>